+++ index.html
@@ -7,23 +7,19 @@
     <script type="module" src="./src/js/components/pages/PurchaseProduct.js"></script>
     <script type="module" src="./src/js/components/pages/ProductManage.js"></script>
     <script type="module" src="./src/js/components/pages/MoneyCharge.js"></script>
-    <script type="module" src="./src/js/components/pages/AddProduct.js"></script>
+    <script type="module" src="./src/js/components/AddProduct.js"></script>
+    <script type="module" src="./src/js/components/ProductList.js"></script>
+    <script type="module" src="./src/js/components/AddCoin.js"></script>
+    <script type="module" src="./src/js/components/CoinList.js"></script>
+    <script type="module" src="./src/js/components/InputMoney.js"></script>
+    <script type="module" src="./src/js/components/PurchaseList.js"></script>
+    <script type="module" src="./src/js/components/LeftCoins.js"></script>
     <link rel="stylesheet" href="./src/css/index.css" />
     <script type="module" src="./index.js"></script>
     <title>NEXTSTEP JavaScript Vending machine</title>
   </head>
 
   <body>
-    <!-- <button id="product-manage-menu">상품 관리</button>
-    <button id="vending-machine-manage-menu">잔돈충전</button>
-    <button id="product-purchase-menu">상품 구매</button> -->
-    <!-- <ul class="tabs">
-      <li class="btn pressed" id="manage-products">상품 관리</li>
-      <li class="btn" id="charging-money">잔돈 충전</li>
-      <li class="btn" id="purchase-product">상품 구매</li>
-    </ul>
-    <div id="app"></div> -->
-
     <tab-menu>
       <ul class="tabs" slot="tab-list">
         <li class="btn pressed" id="product-manage-menu">상품 관리</li>
