+++ src/js/components/pages/PurchaseProduct.js
@@ -12,14 +12,11 @@ export default class PurchaseProduct extends HTMLElement {
   init() {
     this.template.innerHTML = `
 			<link rel="stylesheet" href="./src/css/index.css" />
-      <div class="purchase-container">
-        <h3>충전하기</h3>
-        <div class="vending-machine-wrapper">
-          <input type="number" name="charge-amount" id="charge-input" />
-          <button id="charge-button">충전하기</button>
-        </div>
-        <p>충전 금액: <span id="charge-amount">0</span>원</p>
-      </div>`;
+			<section data-cy="purchase-product">
+				<input-money></input-money>
+				<purchase-list></purchase-list>
+				<left-coins></left-coins>
+		</section>`;
     this.shadow.appendChild(this.template.content.cloneNode(true));
   }
 }
